% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{get_final_annual_sim_all}
\alias{get_final_annual_sim_all}
\title{get_final_annual_sim_all helper function to retrieve powerspectrum.}
\usage{
get_final_annual_sim_all(annual_prcp, years, num_year_sim, num_trials = 1,
  num_autocor_annual_changes = 1, num_magnitude_noise_annual_changes = 1)
}
\arguments{
\item{fig_dir}{Directory to create PDF containing the plot}

\item{PowerSpectrum, }{create this data this using Final_Annual_Sim_All and with its output getLowFrequencyVariability}

\item{ftag}{string to appear in filename}
}
\description{
get_final_annual_sim_all helper function to retrieve powerspectrum.
}
\examples{
ar_models <- list()
climate_variable <- annual_prcp
get_final_annual_sim_all(num_autocor_annual_changes=c(1),num_magnitude_noise_annual_changes=c(1), ar_models, new_ar_model)


}

