% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{getCorrelations}
\alias{getCorrelations}
\title{getCorrelations calculate correlations for plot function 5}
\usage{
getCorrelations(AnnualMeans, vars = c("prcp", "tmax", "tmin"))
}
\arguments{
\item{AnnualMeans}{This value can be retrieved from getObservedStatistics(...)$AnnualMeans or getSimulatedStatistics(..)$AnnualMeans}
}
\description{
getCorrelations calculate correlations for plot function 5
}
\examples{
myobsStats$cors <- getCorrelations(AnnualMeans=myobsStats$AnnualMeans[-1, -1, , , drop=FALSE])


}

