% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{getLowFrequencyVariability}
\alias{getLowFrequencyVariability}
\title{getCorrelations calculate correlations for plot function 5}
\usage{
getLowFrequencyVariability(ANNUAL_PRCP, years, Final_Annual_Sim_All)
}
\arguments{
\item{AnnualMeans}{This value can be retrieved from getObservedStatistics(...)$AnnualMeans or getSimulatedStatistics(..)$AnnualMeans}
}
\description{
getCorrelations calculate correlations for plot function 5
}
\examples{
myobsStats$cors <- getCorrelations(AnnualMeans=myobsStats$AnnualMeans[-1, -1, , , drop=FALSE])



}

