% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{getObservedStatistics}
\alias{getObservedStatistics}
\title{Generate Observation Statistics from source data}
\usage{
getObservedStatistics(wgen_out_array, vars = c("prcp", "tmax", "tmin"),
  timesteps = c("daily", "monthly", "annual"), stats = c("mean", "stdev",
  "skew"), spells = c("wet", "dry"), return.periods = c(10, 30, 100),
  mo = c(1:12))
}
\arguments{
\item{wgen_out_array, }{2 dimensional array of dataframes as collected of returns my wgen_daily, rows are sites, cols are trials}

\item{vars}{usually keep on default c("prcp","tmax","tmin"). Only change if you know what you do.}

\item{timesteps}{usually keep on default c("daily","monthly", "annual"). Only change if you know what you do.}

\item{stats}{usually keep on default c("mean","stdev","skew"). Only change if you know what you do.}

\item{year}{contains the year of the observed data}

\item{Month}{contains the year of the observed data}
}
\description{
Generate Observation Statistics from source data
}
\examples{
getObservedStatistics3(wgen_out_array)
 

}

