% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{getSimulatedStatistics}
\alias{getSimulatedStatistics}
\title{Generate Simulated Statistics from calculated data}
\usage{
getSimulatedStatistics(wgen_out_array, stats = c("mean", "stdev", "skew"),
  vars = c("prcp", "tmax", "tmin"), timesteps = c("daily", "monthly",
  "annual"), mo = c(1:12), return.periods = c(10, 30, 100),
  spells = c("wet", "dry"))
}
\arguments{
\item{wgen_out_array}{outputs of wgen_daily per site and trials collected in array.}

\item{stats}{c("mean","stdev","skew")}

\item{vars}{usually keep on default c("prcp","tmax","tmin"). Only change if you know what you do.}

\item{timesteps}{usually keep on default c("daily","monthly", "annual"). Only change if you know what you do.}

\item{mo}{usually keep on default c(1:12). Only change if you know what you do.}

\item{return.periods}{usually keep on default c(10,30,100). Only change if you know what you do.}

\item{s??ells}{usually keep on default c("wet","dry"). Only change if you know what you do.}
}
\description{
Generate Simulated Statistics from calculated data
}
\examples{
getSimulatedStatistics(sites_prcp, sites_tmax, sites_tmin, num_trials=1, num_site=3, as.matrix(format(sim$out["DATE"], "\%Y")),  as.matrix(sim$out["MONTH"]), vars=c("prcp","tmax","tmin"), timesteps= c("daily","monthly", "annual"), stats=c("mean","stdev","skew"))
 
}

