% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_basinwide_extreme_annual}
\alias{plot_basinwide_extreme_annual}
\title{Plot annual extremes}
\usage{
plot_basinwide_extreme_annual(obs, sim, fig_dir = NULL, ftag = "",
  vars = c("prcp", "tmax", "tmin"), return.periods = c(10, 30, 100))
}
\arguments{
\item{obs}{observation data}

\item{sim}{simulation output}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}
}
\description{
Plot annual extremes
}
\examples{
plot_basinwide_extreme_annual(obs, sim)


}

