% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_basinwide_powerspectrum}
\alias{plot_basinwide_powerspectrum}
\title{Plot PowerSpectrum}
\usage{
plot_basinwide_powerspectrum(PowerSpectrum, fig_dir = NULL, ftag = "")
}
\arguments{
\item{PowerSpectrum, }{create this data this using Final_Annual_Sim_All and with its output getLowFrequencyVariability}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}
}
\description{
Plot PowerSpectrum
}
\examples{
plot7.basinwide.PowerSpectruml (fig_dir, PowerSpectrum)


}

