% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_basinwide_spells_monthly}
\alias{plot_basinwide_spells_monthly}
\title{Plot Monthly spells}
\usage{
plot_basinwide_spells_monthly(obs, sim, fig_dir = NULL, ftag = "",
  stats = c("mean", "stdev", "skew"), vars = c("prcp", "tmax", "tmin"),
  mo = c(1:12), spells = c("wet", "dry"))
}
\arguments{
\item{obs}{observation data}

\item{sim}{simulation output}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}
}
\description{
Plot Monthly spells
}
\examples{
plot_basinwide_spells_monthly(fig_dir, obs, sim)


}

