% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_basinwide_vars_monthly}
\alias{plot_basinwide_vars_monthly}
\title{Plot Monthly variables}
\usage{
plot_basinwide_vars_monthly(obs, sim, ftag = "", fig_dir = NULL,
  stats = c("mean", "stdev", "skew"), vars = c("prcp", "tmax", "tmin"),
  mo = c(1:12))
}
\arguments{
\item{obs}{observation data}

\item{sim}{simulation output}

\item{ftag}{string to appear in filename}

\item{fig_dir}{Directory to create PDF containing the plot. Can be NA to plot on screen}
}
\description{
Plot Monthly variables
}
\examples{
plot_basinwide_vars_monthly(fig_dir, obs, sim)


}

