% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_sitespecific_correlations_annual}
\alias{plot_sitespecific_correlations_annual}
\title{Plot sitespecific annual correlations}
\usage{
plot_sitespecific_correlations_annual(obsCors, simCors, fig_dir = NULL,
  ftag = "Validation", vars = c("prcp", "tmax", "tmin"))
}
\arguments{
\item{obsCors}{observation data}

\item{simCors}{simulation output}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}

\item{vars}{variables}
}
\description{
Plot sitespecific annual correlations
}
\examples{
plot_sitespecific_correlations_annual (obsCors, simCors)


}

