% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_sitespecific_spells_annual}
\alias{plot_sitespecific_spells_annual}
\title{Plot sitespecific annual spells}
\usage{
plot_sitespecific_spells_annual(obs, sim, fig_dir = NULL,
  ftag = "Validation", stats = c("mean", "stdev", "skew"),
  vars = c("prcp", "tmax", "tmin"), mo = c(1:12), spells = c("wet",
  "dry"))
}
\arguments{
\item{obs}{observation data}

\item{sim}{simulation output data}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}
}
\description{
Plot sitespecific annual spells
}
\examples{
plot_sitespecific_spells_annual (obs, sim)


}

