% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_sitespecific_vars}
\alias{plot_sitespecific_vars}
\title{Plot sitespecific variables}
\usage{
plot_sitespecific_vars(obs, sim, fig_dir = NULL, ftag = "",
  timestep = "daily", stats = c("mean", "stdev", "skew"), vars = c("prcp",
  "tmax", "tmin"), mo = c(1:12))
}
\arguments{
\item{obs}{observation data}

\item{sim}{simulation output}

\item{fig_dir}{Directory to create PDF containing the plot}

\item{ftag}{string to appear in filename}
}
\description{
Plot sitespecific variables
}
\examples{
plot_sitespecific_vars(obs, sim)


}

